requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
pydantic-ai
python-dotenv
asyncpg
jinja2
requests
